requirements.txt
Pillow==8.0.1
docopts==0.6.1
progress==1.5
